WIP - modifying Auth and Toast context to be easier to use and moved it. Organizing client folder structure
