+++ client/src/main.tsx
@@ -2,9 +2,14 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
+import { AuthProvider, ToastProvider } from './components/utils/Context.tsx';
 
 createRoot(document.getElementById('root')!).render(
 	<StrictMode>
-		<App />
+		<AuthProvider>
+			<ToastProvider>
+				<App />
+			</ToastProvider>
+		</AuthProvider>
 	</StrictMode>
 );
